Hopefully finally fixed TWEAKed attributes and Tool
Whoever thought of allowing the same attribute to either be a single
object, or an array of other objects, should be talked to severely.

Also TIL that Nil ~~ Cool.

AnyTool is gone again: the onus of handling a "tools" atribute that
can both be a scalar, as well as an array, is now the classes that
have that (notably: Metadata and Component).

*phew*